(Garbage collector) .apk file
 Garbage collector Android project is a mobile application built on the Android Studio platform that incorporates Firebase connectivity to manage user authentication and data. The app is designed to make garbage collection more efficient and convenient for users.

One of the major features of the app is its live location support, which allows users to request garbage collection from their current location. The app uses GPS technology to pinpoint the user's location, making it easy for them to request garbage pickup without having to provide their address or other location details manually and it completely built upon realtime database which makes updation to the application rapidly fast.

Location can only be accessed unless and untill user wishes too

In addition to requesting garbage pickup, the app also allows users to create customer complaints. This feature enables users to report any issues they encounter with the garbage collection process, such as missed pickups or improper disposal. The complaints are then sent to the relevant authorities for action.

Other users of the application can also view the compliaints if they want

Another useful feature of the app is the ability to notify the administrator when specific bins are full. This helps to ensure that garbage collection is done on time and that overflowing bins are emptied promptly.

Overall, this garbage collector Android application is a powerful tool that streamlines the garbage collection process, making it more efficient and convenient for users. Its robust features, such as live location support, customer complaints, and bin notifications, make it a must-have for anyone looking to improve their garbage collection process.